requirements.txt
Django>=5.2.0
Pillow>=8.0.0
pytz>=2021.1

# Facial recognition dependencies
tensorflow>=2.5.0
mtcnn>=0.1.0
keras-vggface>=0.6
opencv-python>=4.5.0
scipy>=1.7.0
numpy>=1.19.0
